Move update, collision and logic to a class character
